Move ApplicationForm previews to subfolder
